Auswertung aufgeteilt
+ Auswertung_kumuliertController
+ Auswertung_EinzelController

• Admin-Home-View
• Home-View (normal user)

• Partial Views for Auswertung split in 2 folders (Auswertung_Einzel and Auswertung_kumuliert)
