+++ Umfrage-Tool/Umfrage-Tool/Controllers/Auswertung_EinzelController.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web.Mvc;
@@ -12,7 +12,7 @@ namespace Umfrage_Tool.Controllers
 {
     [Authorize(Roles = "Ersteller, Admin")]
 
-    public class Umfrage_ErgebnisseController : Controller
+    public class Auswertung_EinzelController : Controller
     {
 
         DatabaseContent db = new DatabaseContent();
@@ -36,9 +36,9 @@ namespace Umfrage_Tool.Controllers
         }
 
 
-        public Umfrage_ErgebnisseController()  { }
+        public Auswertung_EinzelController() { }
 
-        public Umfrage_ErgebnisseController(ApplicationUserManager userManager)
+        public Auswertung_EinzelController(ApplicationUserManager userManager)
         {
             UserManager = userManager;
         }
@@ -54,7 +54,7 @@ namespace Umfrage_Tool.Controllers
                 .FirstOrDefault(t => t.ID == umfrageId);
             ICollection<SessionViewModel> sessionListe = _sessionZuViewTransformer.ListTransform(ausgewählteUmfrage?.sessions).ToList();
             sessionListe = sessionListe.OrderByDescending(m => m.creationDate).ToList();
-            sessionListe.First().surveyviewModel = 
+            sessionListe.First().surveyviewModel =
                 _umfrageZuViewTransformerMitFragen.Transform(ausgewählteUmfrage);
 
             if (ausgewählteUmfrage == null || (!BenutzerDarfDas(ausgewählteUmfrage.Creator) || ausgewählteUmfrage.states != Survey.States.Beendet))
@@ -118,33 +118,6 @@ namespace Umfrage_Tool.Controllers
             return PartialView(beantwortung);
         }
 
-        public ActionResult Fragen_Ergebnisse(Guid arg)
-        {
-            Guid umfrageId = arg;
-            Survey ausgewählteUmfrage = db.Surveys
-                .Include(a => a.questions
-                .Select(c => c.givenAnswer.Select(k => k.session)))
-                .Include(s => s.questions
-                .Select(t => t.choice))
-                .FirstOrDefault(b => b.ID == umfrageId);
-            var fragenListe = _fragenZuViewTransformer.ListTransform(ausgewählteUmfrage?.questions);
-            fragenListe = fragenListe.OrderBy(u => u.position).ToList();
-
-            fragenListe.First().surveyViewModel =
-                _umfrageZuViewTransformerMitFragen.Transform(ausgewählteUmfrage);
-
-            if (!BenutzerDarfDas(fragenListe.First().surveyViewModel.Creator) || fragenListe.First().surveyViewModel.states != Survey.States.Beendet)
-            {
-                return RedirectToAction("StatusUmfrageAuswertung", "Fehlermeldungen");
-            }
-            if (db.Surveys.First(s => s.ID == arg).sessions == null) 
-            {
-                return RedirectToAction("AuswertungKeineAntworten", "Fehlermeldungen");
-            }
-
-            return View(fragenListe.ToList());
-        }
-
         private bool BenutzerDarfDas(Guid creator)
         {
             var benutzerText = UserManager.Users.First(f => f.Id == creator.ToString()).Email;
@@ -152,35 +125,6 @@ namespace Umfrage_Tool.Controllers
             var darfErDasWirklich = a == benutzerText || User.Identity.Name == "Admin@FI17.de";
 
             return darfErDasWirklich;
-        }
-
-        public PartialViewResult Panel_fuer_Frage_in_kumulierter_Auswertung(QuestionViewModel frage)
-        {
-            return PartialView(frage);
-        }
-        public PartialViewResult Freitext_Kumuliert(QuestionViewModel frage)
-        {
-            return PartialView(frage);
-        }
-        public PartialViewResult MultipleOne_Kumuliert(QuestionViewModel frage)
-        {
-            return PartialView(frage);
-        }
-        public PartialViewResult Skalenfrage_Kumuliert(QuestionViewModel frage)
-        {
-            return PartialView(frage);
-        }
-        public PartialViewResult MultipleOneMitSonstiges_Kumuliert(QuestionViewModel frage)
-        {
-            return PartialView(frage);
-        }
-        public PartialViewResult MultipleMore_Kumuliert(QuestionViewModel frage)
-        {
-            return PartialView(frage);
-        }
-        public PartialViewResult MultipleMoreMitSonstiges_Kumuliert(QuestionViewModel frage)
-        {
-            return PartialView(frage);
         }
     }
 }
