+++ Umfrage-Tool/Umfrage-Tool/Umfrage-Tool.csproj
@@ -168,11 +168,12 @@
     <Compile Include="App_Start\RouteConfig.cs" />
     <Compile Include="App_Start\Startup.Auth.cs" />
     <Compile Include="Controllers\AccountController.cs" />
+    <Compile Include="Controllers\Auswertung_EinzelController.cs" />
     <Compile Include="Controllers\FehlermeldungenController.cs" />
     <Compile Include="Controllers\Umfrage_BeantwortungController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Controllers\ManageController.cs" />
-    <Compile Include="Controllers\Umfrage_ErgebnisseController.cs" />
+    <Compile Include="Controllers\Auswertung_kumuliertController.cs" />
     <Compile Include="Controllers\Umfrage_ErstellungController.cs" />
     <Compile Include="database.cs" />
     <Compile Include="FromModelTransformer\Answering\ModelToAnsweringTransformer.cs" />
@@ -278,30 +279,30 @@
     <Content Include="Views\Umfrage_Beantwortung\Freitext.cshtml" />
     <Content Include="Views\Umfrage_Erstellung\FrageErstellung.cshtml" />
     <Content Include="Views\Umfrage_Beantwortung\Umfrage_beendet.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\Antworten.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\Ergebnisse.cshtml" />
+    <Content Include="Views\Auswertung_Einzel\Antworten.cshtml" />
+    <Content Include="Views\Auswertung_Einzel\Ergebnisse.cshtml" />
     <Content Include="Views\Umfrage_Erstellung\Vorschau.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\Fragen_Ergebnisse.cshtml" />
+    <Content Include="Views\Auswertung_kumuliert\Fragen_Ergebnisse.cshtml" />
     <Content Include="Views\Umfrage_Beantwortung\MultipleOne.cshtml" />
     <Content Include="Views\Umfrage_Beantwortung\Skalenfrage.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\Freitext_Kumuliert.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\MultipleOne_Kumuliert.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\Skalenfrage_Kumuliert.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\Panel_fuer_Frage_in_kumulierter_Auswertung.cshtml" />
+    <Content Include="Views\Auswertung_kumuliert\Freitext_Kumuliert.cshtml" />
+    <Content Include="Views\Auswertung_kumuliert\MultipleOne_Kumuliert.cshtml" />
+    <Content Include="Views\Auswertung_kumuliert\Skalenfrage_Kumuliert.cshtml" />
+    <Content Include="Views\Auswertung_kumuliert\Panel_fuer_Frage_in_kumulierter_Auswertung.cshtml" />
     <Content Include="Views\Umfrage_Erstellung\Skalenfragen_Erstellung.cshtml" />
     <Content Include="Views\Umfrage_Erstellung\Multiple_Choice_Fragen_Erstellung.cshtml" />
     <Content Include="Views\Umfrage_Beantwortung\MultipleMore.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\Freitext_Einzel.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\Skala_Einzel.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\MultipleOne_Einzel.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\MultipleMore_Einzel.cshtml" />
+    <Content Include="Views\Auswertung_Einzel\Freitext_Einzel.cshtml" />
+    <Content Include="Views\Auswertung_Einzel\Skala_Einzel.cshtml" />
+    <Content Include="Views\Auswertung_Einzel\MultipleOne_Einzel.cshtml" />
+    <Content Include="Views\Auswertung_Einzel\MultipleMore_Einzel.cshtml" />
     <Content Include="Views\Umfrage_Beantwortung\MultipleOneMitSonstiges.cshtml" />
     <Content Include="Views\Umfrage_Beantwortung\MultipleMoreMitSonstiges.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\MultipleOneMitSonstiges_Kumuliert.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\MultipleMore_Kumuliert.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\MultipleMoreMitSonstiges_Kumuliert.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\MultipleOneMitSonstiges_Einzel.cshtml" />
-    <Content Include="Views\Umfrage_Ergebnisse\MultipleMoreMitSonstiges_Einzel.cshtml" />
+    <Content Include="Views\Auswertung_kumuliert\MultipleOneMitSonstiges_Kumuliert.cshtml" />
+    <Content Include="Views\Auswertung_kumuliert\MultipleMore_Kumuliert.cshtml" />
+    <Content Include="Views\Auswertung_kumuliert\MultipleMoreMitSonstiges_Kumuliert.cshtml" />
+    <Content Include="Views\Auswertung_Einzel\MultipleOneMitSonstiges_Einzel.cshtml" />
+    <Content Include="Views\Auswertung_Einzel\MultipleMoreMitSonstiges_Einzel.cshtml" />
     <Content Include="Views\Home\Erstellte_Umfragen_Normaler_Nutzer.cshtml" />
     <Content Include="Views\Home\Erstellte_Umfragen_Admin.cshtml" />
     <Content Include="Views\Umfrage_Erstellung\Freitext.cshtml" />
