+++ Umfrage-Tool/Umfrage-Tool/Views/Auswertung_Einzel/Antworten.cshtml
@@ -32,29 +32,29 @@
                     switch (Beantwortung.questionViewModel.type)
                     {
                         case Domain.Question.choices.MultipleOne:
-                            @Html.Action("MultipleOne_Einzel", "Umfrage_Ergebnisse", Beantwortung);
+                            @Html.Action("MultipleOne_Einzel", "Auswertung_Einzel", Beantwortung);
                             frage_wurde_geschrieben[Beantwortung.questionViewModel.ID.ToString()] = true;
                             break;
                         case Domain.Question.choices.Skalenfrage:
-                            @Html.Action("Skala_Einzel", "Umfrage_Ergebnisse", Beantwortung);
+                            @Html.Action("Skala_Einzel", "Auswertung_Einzel", Beantwortung);
                             frage_wurde_geschrieben[Beantwortung.questionViewModel.ID.ToString()] = true;
                             break;
                         case Domain.Question.choices.MultipleMore:
-                            @Html.Action("MultipleMore_Einzel", "Umfrage_Ergebnisse", Beantwortung);
+                            @Html.Action("MultipleMore_Einzel", "Auswertung_Einzel", Beantwortung);
                             frage_wurde_geschrieben[Beantwortung.questionViewModel.ID.ToString()] = true;
                             break;
                         case Domain.Question.choices.MultipleOneMitSonstiges:
-                            @Html.Action("MultipleOneMitSonstiges_Einzel", "Umfrage_Ergebnisse", Beantwortung);
+                            @Html.Action("MultipleOneMitSonstiges_Einzel", "Auswertung_Einzel", Beantwortung);
                             frage_wurde_geschrieben[Beantwortung.questionViewModel.ID.ToString()] = true;
                             break;
                         case Domain.Question.choices.MultipleMoreMitSonstiges:
-                            @Html.Action("MultipleMoreMitSonstiges_Einzel", "Umfrage_Ergebnisse", Beantwortung);
+                            @Html.Action("MultipleMoreMitSonstiges_Einzel", "Auswertung_Einzel", Beantwortung);
                             frage_wurde_geschrieben[Beantwortung.questionViewModel.ID.ToString()] = true;
                             break;
 
                         case Domain.Question.choices.Freitext:
                         default:
-                            @Html.Action("Freitext_Einzel", "Umfrage_Ergebnisse", Beantwortung);
+                            @Html.Action("Freitext_Einzel", "Auswertung_Einzel", Beantwortung);
                             frage_wurde_geschrieben[Beantwortung.questionViewModel.ID.ToString()] = true;
                             break;
                     }
@@ -63,7 +63,7 @@
 
         </div>
         <div id="PanelFooter" class="panel-footer">
-            @Html.ActionLink("Zurück", "Ergebnisse", "Umfrage_Ergebnisse", new { arg = Session["Vorherige_Umfrage"].ToString() }, new { @class = "btn btn-lg btn-primary KnöpfeFürBeantwortung" })
+            @Html.ActionLink("Zurück", "Ergebnisse", "Auswertung_Einzel", new { arg = Session["Vorherige_Umfrage"].ToString() }, new { @class = "btn btn-lg btn-primary KnöpfeFürBeantwortung" })
         </div>
     </div>
 </div>
