+++ Umfrage-Tool/Umfrage-Tool/Views/Auswertung_Einzel/Ergebnisse.cshtml
@@ -37,7 +37,7 @@
                         <td>
                             @if (beantwortungssitzung.givenAnswerViewModels.Count != 0)
                             {
-                                @Html.ActionLink("Anzeigen", "Antworten", "Umfrage_Ergebnisse", new { arg = beantwortungssitzung.ID }, new { @class = "btn btn-info btn-sm" })
+                                @Html.ActionLink("Anzeigen", "Antworten", "Auswertung_Einzel", new { arg = beantwortungssitzung.ID }, new { @class = "btn btn-info btn-sm" })
                             }
                             else
                             {
@@ -50,6 +50,6 @@
         </table>
     </div>
     <div class="panel-footer">
-        @Html.ActionLink("Zurück", "Fragen_Ergebnisse", "Umfrage_Ergebnisse",  new { arg = Model.First().surveyviewModel.ID },new { @class = "btn btn-info btn-md" })
+        @Html.ActionLink("Zurück", "Fragen_Ergebnisse", "Auswertung_kumuliert",  new { arg = Model.First().surveyviewModel.ID },new { @class = "btn btn-info btn-md" })
     </div>
 </div>
