+++ Umfrage-Tool/Umfrage-Tool/Views/Auswertung_Einzel/Freitext_Einzel.cshtml
