+++ Umfrage-Tool/Umfrage-Tool/Views/Auswertung_Einzel/MultipleMore_Einzel.cshtml
