+++ Umfrage-Tool/Umfrage-Tool/Views/Auswertung_Einzel/Skala_Einzel.cshtml
