+++ Umfrage-Tool/Umfrage-Tool/Views/Auswertung_kumuliert/Fragen_Ergebnisse.cshtml
@@ -22,10 +22,10 @@
     <div class="panel-group" id="hauptpanelFürAutomatischesEinklappen">
         @foreach (var frage in Model)
         {
-            @Html.Action("Panel_fuer_Frage_in_kumulierter_Auswertung", "Umfrage_Ergebnisse", frage)
+            @Html.Action("Panel_fuer_Frage_in_kumulierter_Auswertung", "Auswertung_kumuliert", frage)
         }
     </div>
-    @Html.ActionLink("Einzelauswertung", "Ergebnisse", "Umfrage_Ergebnisse", new { arg = Model.First().surveyViewModel.ID }, new { @class = "btn btn-info" })
+@Html.ActionLink("Einzelauswertung", "Ergebnisse", "Auswertung_Einzel", new { arg = Model.First().surveyViewModel.ID }, new { @class = "btn btn-info" })
 <br />
 <br />
 @Html.ActionLink("Zurück", "Index", "Home", new { area = "" }, new { @class = "btn btn-info" })
