+++ Umfrage-Tool/Umfrage-Tool/Views/Auswertung_kumuliert/MultipleOne_Kumuliert.cshtml
