+++ Umfrage-Tool/Umfrage-Tool/Views/Home/Erstellte_Umfragen_Admin.cshtml
@@ -82,7 +82,7 @@
                             </script>
                         </td>
                         <td>
-                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Umfrage_Ergebnisse", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
+                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Auswertung_kumuliert", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
                         </td>
                         <td>
                             @Html.ActionLink("bearbeiten", "FrageErstellung", "Umfrage_Erstellung", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs" })
@@ -117,7 +117,7 @@
                             </script>
                         </td>
                         <td>
-                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Umfrage_Ergebnisse", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
+                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Auswertung_kumuliert", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
                         </td>
                         <td>
                             @Html.ActionLink("bearbeiten", "FrageErstellung", "Umfrage_Erstellung", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
@@ -152,7 +152,7 @@
                             </script>
                         </td>
                         <td>
-                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Umfrage_Ergebnisse", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs" })
+                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Auswertung_kumuliert", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs" })
                         </td>
                         <td>
                             @Html.ActionLink("bearbeiten", "FrageErstellung", "Umfrage_Erstellung", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
