+++ Umfrage-Tool/Umfrage-Tool/Views/Home/Erstellte_Umfragen_Normaler_Nutzer.cshtml
@@ -71,7 +71,7 @@
                             <button type="button" class="btn btn-info btn-xs"  onclick="copyLink('myInput_@Umfrage.ID')">Link kopieren</button>
                         </td>
                         <td>
-                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Umfrage_Ergebnisse", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
+                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Auswertung_kumuliert", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
                         </td>
                         <td>
                             @Html.ActionLink("bearbeiten", "FrageErstellung", "Umfrage_Erstellung", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs" })
@@ -98,7 +98,7 @@
                             <button type="button" class="btn btn-info btn-xs" onclick="copyLink('myInput_@Umfrage.ID')">Link kopieren</button>
                         </td>
                         <td>
-                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Umfrage_Ergebnisse", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
+                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Auswertung_kumuliert", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
                         </td>
                         <td>
                             @Html.ActionLink("bearbeiten", "FrageErstellung", "Umfrage_Erstellung", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
@@ -125,7 +125,7 @@
                             <button type="button" class="btn btn-info btn-xs"  onclick="copyLink('myInput_@Umfrage.ID')">Link kopieren</button>
                         </td>
                         <td>
-                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Umfrage_Ergebnisse", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs" })
+                            @Html.ActionLink("auswerten", "Fragen_Ergebnisse", "Auswertung_kumuliert", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs" })
                         </td>
                         <td>
                             @Html.ActionLink("bearbeiten", "FrageErstellung", "Umfrage_Erstellung", new { arg = Umfrage.ID }, new { @class = "btn btn-info btn-xs", disabled = "disabled" })
